BLAS Traversal struggles
I rewrote most of the BLAS function so it uploads the data to the GPU but I'm getting a really weird warning that it's accessing an std::vector which can't be the case.

I also removed a few CPU only vectors. I plan to remove all of them and see how it'll work.